Adding the draft and still in work portfolio
This is still a work in progress.  There have been some visuals loaded on the gamma portfolio, that I have yet to add my annotations.  I have a lot to say, so Hope to update soon.  As well, I feel like I am just getting started with my analysis on this big dataset.  It has been fun so far.